# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Avalanche P-Chain and EVM Indexer API\n\n## Overview\nThis API provides access to indexed Avalanche blockchain data including:\n- **Data API** (`/api/v1/data/*`): Blocks, transactions, subnets, validators, P-chain data\n- **Metrics API** (`/api/v1/metrics/*`): Fee statistics, chain metrics, time series data\n\n## Rate Limiting\nAll endpoints are rate limited to 60 requests/minute per IP with a burst of 10.",
        "title": "Icicle API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/data/validators": {
            "get": {
                "description": "Get a paginated list of L1 validators with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - Validators"
                ],
                "summary": "List validators",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by subnet ID",
                        "name": "subnet_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by active status (true/false)",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.Validator"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/api.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/data/validators/{id}": {
            "get": {
                "description": "Get details for a specific validator by validation ID or node ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - Validators"
                ],
                "summary": "Get validator by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Validation ID or Node ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.Validator"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/data/validators/{id}/deposits": {
            "get": {
                "description": "Get balance transactions (deposits, topups, refunds) for a validator",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - Validators"
                ],
                "summary": "Get validator deposits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Validation ID or Node ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.ValidatorDeposit"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/api.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ErrorCode"
                },
                "details": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "retry_after": {
                    "type": "integer"
                }
            }
        },
...
        "api.ErrorCode": {
            "type": "string",
            "enum": [
                "INVALID_PARAMETER",
                "NOT_FOUND",
                "INTERNAL_ERROR",
                "RATE_LIMITED",
                "VALIDATION_FAILED",
                "DATABASE_ERROR"
            ],
            "x-enum-varnames": [
                "ErrInvalidParameter",
                "ErrNotFound",
                "ErrInternalError",
                "ErrRateLimited",
                "ErrValidationFailed",
                "ErrDatabaseError"
            ]
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.APIError"
                }
            }
        },
...
        "api.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/api.Meta"
                }
            }
        },
...
        "api.Validator": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "balance": {
                    "type": "integer",
                    "example": 100000000000
                },
                "end_time": {
                    "type": "string"
                },
                "fees_paid": {
                    "type": "integer",
                    "example": 5000000000
                },
                "initial_deposit": {
                    "type": "integer",
                    "example": 100000000000
                },
                "node_id": {
                    "type": "string",
                    "example": "NodeID-P7oB2McjBGgW..."
                },
                "refund_amount": {
                    "type": "integer",
                    "example": 0
                },
                "start_time": {
                    "type": "string"
                },
                "subnet_id": {
                    "type": "string",
                    "example": "2XDnKyAEr..."
                },
                "total_topups": {
                    "type": "integer",
                    "example": 50000000000
                },
                "uptime_percentage": {
                    "type": "number",
                    "example": 99.5
                },
                "validation_id": {
                    "type": "string",
                    "example": "2ZW6HUePB..."
                },
                "weight": {
                    "type": "integer",
                    "example": 2000
                }
            }
        },
        "api.ValidatorDeposit": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 10000000000
                },
                "block_number": {
                    "type": "integer",
                    "example": 12345678
                },
                "block_time": {
                    "type": "string"
                },
                "tx_id": {
                    "type": "string",
                    "example": "2ZW6HUePB..."
                },
                "tx_type": {
                    "type": "string",
                    "example": "IncreaseL1ValidatorBalanceTx"
                }
            }
        }
    },
...
}